[MI-2230] Move record types to constants file
